+++ scripts/lpp_extension.py
@@ -22,6 +22,52 @@ def set_no_config(*args):
         setattr(control, "do_not_save_to_config", True)
 
 
+def try_send_derpi_request(lpp, *args):
+    try:
+        lpp.send_derpibooru_request(*args)
+        return format_status_msg(
+            lpp, "Successfully fetched tags from Derpibooru."
+        )
+    except Exception as e:
+        return format_status_msg(
+            lpp, f"Filed to fetch tags: {str(e)}"
+        )
+
+
+def try_save_prompts(lpp, name):
+    try:
+        lpp.cache_current_prompts(name)
+        return format_status_msg(
+            lpp, f"Prompts saved as \"{name}\"."
+        )
+    except Exception as e:
+        return format_status_msg(
+            lpp, f"Failed to save prompts: {str(e)}."
+        )
+
+
+def try_load_prompts(lpp, name):
+    try:
+        lpp.load_cached_prompts(name)
+        return format_status_msg(lpp, f"Loaded \"{name}\".")
+    except Exception as e:
+        return format_status_msg(
+            lpp, f"Failed to load prompts: {str(e)}."
+        )
+
+
+def try_delete_prompts(lpp, name):
+    try:
+        lpp.delete_cached_prompts(name)
+        return format_status_msg(
+            lpp, f"\"{name}\" deleted."
+        )
+    except Exception as e:
+        return format_status_msg(
+            lpp, f"Failed to delete prompts: {str(e)}."
+        )
+
+
 class Scripts(scripts.Script):
     def __init__(self):
         self.lpp = LPP(base_dir)
@@ -47,7 +93,6 @@ class Scripts(scripts.Script):
                     label="Include Standard Negative Prompt",
                     value=self.config["include_standard_negative_prompt"]
                 )
-
             with gr.Column(variant="compact"):
                 # Derpibooru Query Panel --------------------------------------
                 with gr.Accordion(
@@ -108,28 +153,29 @@ class Scripts(scripts.Script):
                         )
 
                 # Save/Load Prompts Panel -------------------------------------
-                with gr.Accordion("Saving & Loading", open=False):
+                with gr.Accordion("Prompts Manager", open=False):
                     with gr.Row():
+                        prompts_manager_input = gr.Dropdown(
+                            label="Prompts Collection Name",
+                            choices=self.lpp.get_cached_prompts_names(),
+                            allow_custom_value=True
+                        )
+                    with gr.Row():
+                        save_prompts_btn = gr.Button(value="Save")
+                        load_prompts_btn = gr.Button(value="Load")
+                        delete_prompts_btn = gr.Button("Delete")
+                    with gr.Row(variant="panel", visible=False) as confirm_action_dialog:
                         with gr.Column():
-                            save_prompts_name = gr.Textbox(
-                                label="Save Current Prompts as"
-                            )
-                            save_prompts_btn = gr.Button(value="Save")
-                        with gr.Column():
-                            load_prompts_name = gr.Dropdown(
-                                label="Load Saved Prompts",
-                                choices=self.lpp.get_cached_prompts_names()
-                            )
                             with gr.Row():
-                                with gr.Column():
-                                    load_prompts_btn = gr.Button(value="Load")
-                                with gr.Column():
-                                    with gr.Row():
-                                        delete_prompts_btn = gr.Button("Delete")
-                                        confirm_delete_prompts_btn = gr.Button("Confirm delete", variant="stop", visible=False)
-                                        cancel_delete_prompts_btn = gr.Button("Cancel", visible=False)
+                                confirm_action_msg = gr.Markdown()
+                                confirm_action_type = gr.Textbox(visible=False)
+                                confirm_action_name = gr.Textbox(visible=False)
+                            with gr.Row():
+                                confirm_action_btn = gr.Button(
+                                    "Confirm", variant="stop")
+                                cancel_action_btn = gr.Button("Cancel")
                     with gr.Row():
-                        load_prompts_metadata = gr.JSON(
+                        prompts_manager_metadata = gr.JSON(
                             label="Prompts Info",
                             show_label=True,
                             visible=False
@@ -146,71 +192,68 @@ class Scripts(scripts.Script):
             set_no_config(enabled, auto_negative_prompt, query_textbox,
                           prompts_count, filter_type, sort_type, prefix,
                           suffix, tag_filter, send_btn, status_bar,
-                          save_prompts_name, load_prompts_name,
+                          prompts_manager_input,
                           save_prompts_btn, load_prompts_btn,
-                          load_prompts_metadata, delete_prompts_btn,
-                          confirm_delete_prompts_btn, cancel_delete_prompts_btn)
+                          prompts_manager_metadata, delete_prompts_btn,
+                          confirm_action_btn, cancel_action_btn,
+                          confirm_action_type, confirm_action_name)
 
             # Event Handlers --------------------------------------------------
             # Send Button Click
-            def send_derpibooru_request(*args, **kwargs):
-                try:
-                    self.lpp.send_derpibooru_request(*args, **kwargs)
-                    return format_status_msg(
-                        self.lpp, "Successfully fetched tags from Derpibooru."
-                    )
-                except Exception as e:
-                    return format_status_msg(
-                        self.lpp, f"Filed to fetch tags: {str(e)}"
-                    )
-
             send_btn.click(
-                lambda *args: send_derpibooru_request(*args),
-                inputs=[query_textbox, prompts_count, filter_type, sort_type],
-                outputs=[status_bar],
+                lambda *args: try_send_derpi_request(self.lpp, *args),
+                [query_textbox, prompts_count, filter_type, sort_type],
+                [status_bar],
                 show_progress="full"
             )
 
             # Save Button Click
-            def save_prompts(name):
-                try:
-                    self.lpp.cache_current_prompts(name)
-                    msg = format_status_msg(
-                        self.lpp, f"Prompts saved as \"{name}\"."
+            def save_prompts_click(name):
+                if name in self.lpp.get_cached_prompts_names():
+                    return (
+                        format_status_msg(self.lpp),
+                        gr.Dropdown.update(
+                            choices=self.lpp.get_cached_prompts_names()
+                        ),
+                        f"Are you sure you want to overwrite \"{name}\"?",
+                        "overwrite",
+                        name,
+                        gr.update(visible=True)
                     )
-                except Exception as e:
-                    msg = format_status_msg(
-                        self.lpp, f"Failed to save prompts: {str(e)}."
+                else:
+                    return (
+                        try_save_prompts(self.lpp, name),
+                        gr.Dropdown.update(
+                            choices=self.lpp.get_cached_prompts_names()
+                        ),
+                        "", "", "", gr.update(visible=False)
                     )
-                return (
-                    msg,
-                    gr.Dropdown.update(
-                        choices=list(self.lpp.get_cached_prompts_names())
-                    ),
-                    gr.Textbox.update(value="")
-                )
 
             save_prompts_btn.click(
-                lambda name: save_prompts(name),
-                inputs=[save_prompts_name],
-                outputs=[status_bar, load_prompts_name, save_prompts_name]
+                save_prompts_click,
+                [prompts_manager_input],
+                [status_bar, prompts_manager_input, confirm_action_msg,
+                 confirm_action_type, confirm_action_name,
+                 confirm_action_dialog]
             )
 
             # Load Button Click
-            def load_prompts(name):
-                try:
-                    self.lpp.load_cached_prompts(name)
-                    msg = format_status_msg(self.lpp, f"Loaded \"{name}\".")
-                except Exception as e:
-                    msg = format_status_msg(
-                        self.lpp, f"Failed to load prompts: {str(e)}."
-                    )
-                return (msg, gr.JSON.update(visible=False))
-
             load_prompts_btn.click(
-                lambda name: load_prompts(name),
-                inputs=[load_prompts_name],
-                outputs=[status_bar, load_prompts_metadata]
+                lambda name: [try_load_prompts(self.lpp, name),
+                              gr.update(visible=False)],
+                [prompts_manager_input],
+                [status_bar, prompts_manager_metadata]
+            )
+
+            # Delete Button Click
+            delete_prompts_btn.click(
+                lambda name: [f"Are you sure you want to delete \"{name}\"?",
+                              "delete",
+                              name,
+                              gr.update(visible=True)],
+                [prompts_manager_input],
+                [confirm_action_msg, confirm_action_type,
+                 confirm_action_name, confirm_action_dialog]
             )
 
             # Load Prompts Dropdown Change
@@ -223,52 +266,39 @@ class Scripts(scripts.Script):
                 except Exception as e:
                     return gr.JSON.update(visible=False)
 
-            load_prompts_name.change(
-                lambda name: load_prompts_metadata_update(name),
-                inputs=[load_prompts_name],
-                outputs=[load_prompts_metadata]
+            prompts_manager_input.change(
+                load_prompts_metadata_update,
+                [prompts_manager_input],
+                [prompts_manager_metadata]
             )
 
-            # Delete Prompts Buttons
-            def delete_prompts(name):
-                try:
-                    self.lpp.delete_cached_prompts(name)
-                    msg = format_status_msg(self.lpp, f"\"{name}\" deleted.")
-                except Exception as e:
-                    msg = format_status_msg(
-                        self.lpp, f"Failed to delete prompts: {str(e)}."
-                    )
+            # Action Confirmation Dialog
+            def invoke_action(name, action_type):
+                if action_type == "delete":
+                    msg = try_delete_prompts(self.lpp, name)
+                    selected_val = ""
+                if action_type == "overwrite":
+                    msg = try_save_prompts(self.lpp, name)
+                    selected_val = name
                 return (
                     msg,
                     gr.Dropdown.update(
                         choices=list(self.lpp.get_cached_prompts_names()),
-                        value=""
+                        value=selected_val
                     ),
-                    gr.update(visible=True),
                     gr.update(visible=False),
                     gr.update(visible=False)
                 )
-            delete_prompts_buttons = [delete_prompts_btn,
-                                      confirm_delete_prompts_btn,
-                                      cancel_delete_prompts_btn]
-            confirm_delete_prompts_btn.click(
-                lambda name: delete_prompts(name),
-                [load_prompts_name],
-                [status_bar, load_prompts_name] + delete_prompts_buttons
+            confirm_action_btn.click(
+                invoke_action,
+                [confirm_action_name, confirm_action_type],
+                [status_bar, prompts_manager_input,
+                 confirm_action_dialog, prompts_manager_metadata]
             )
-            delete_prompts_btn.click(
-                lambda: [gr.update(visible=False),
-                         gr.update(visible=True),
-                         gr.update(visible=True)],
+            cancel_action_btn.click(
+                lambda: gr.update(visible=False),
                 None,
-                delete_prompts_buttons
-            )
-            cancel_delete_prompts_btn.click(
-                lambda: [gr.update(visible=True),
-                         gr.update(visible=False),
-                         gr.update(visible=False)],
-                None,
-                delete_prompts_buttons
+                [confirm_action_dialog]
             )
         return [enabled, auto_negative_prompt, prefix, suffix, tag_filter]